requirements.txt
# Core dependencies
streamlit==1.32.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3

# Vector database
qdrant-client==1.7.0

# Embedding and LLM dependencies
faiss-cpu==1.7.4

# Async support
httpx==0.25.0
asyncio==3.4.3

# Logging and monitoring
loguru==0.7.0

# Configuration management
pydantic==2.4.2
pydantic-settings==2.0.3

# Authentication and security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# PDF processing
pypdf==3.17.1
pdfplumber==0.10.3
pdf2image==1.17.0
pytesseract==0.3.10
tabula-py==2.9.0
camelot-py==0.11.0
opencv-python==4.8.1.78

# Data processing
pandas==2.0.3
matplotlib==3.7.3
plotly==5.18.0

# Image processing
pillow==10.1.0
unstructured==0.10.30
unstructured-inference==0.7.21

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1